Move already exists exceptions to metastore module
