+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/metastore/AbstractTestHiveMetastore.java
@@ -16,9 +16,9 @@ package io.trino.plugin.hive.metastore;
 import io.trino.metastore.Column;
 import io.trino.metastore.Database;
 import io.trino.metastore.HiveMetastore;
+import io.trino.metastore.SchemaAlreadyExistsException;
 import io.trino.metastore.Table;
-import io.trino.plugin.hive.SchemaAlreadyExistsException;
-import io.trino.plugin.hive.TableAlreadyExistsException;
+import io.trino.metastore.TableAlreadyExistsException;
 import io.trino.spi.connector.TableNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
